requirements.txt
matplotlib==3.10.7
numpy==2.3.5
opencv_python==4.12.0.88
opencv_python_headless==4.10.0.84
pandas==2.3.3
Pillow==12.0.0
PyYAML==6.0.1
PyYAML==6.0.3
scikit_learn==1.7.2
seaborn==0.13.2
torch==2.5.1+cu121
torch==2.9.1
torchvision==0.20.1+cu121
torchvision==0.24.1
tqdm==4.66.5
tqdm==4.67.1
transformers==4.57.3
ultralytics==8.3.224
ultralytics==8.3.233
